+++ projects/valtimo/components/src/lib/components/page-header/page-header.component.html
@@ -20,7 +20,8 @@
   [ngClass]="{
     '--hidden': (showPageHeader$ | async) === false,
     '--compact': compactMode$ | async,
-    '--no-content': (compactMode$ | async) && (pageActionsHasContent$ | async) === false
+    '--no-content': (compactMode$ | async) && (pageActionsHasContent$ | async) === false,
+    '--no-margin': pageTitleHidden$ | async
   }"
 >
   <ng-container #contentVcr></ng-container>
