+++ projects/valtimo/components/src/lib/components/page-header/page-header.component.scss
@@ -29,7 +29,8 @@
     margin-bottom: 16px;
   }
 
-  &.--no-content {
+  &.--no-content,
+  &.--no-margin {
     margin-bottom: 0;
   }
 }
