+++ projects/valtimo/components/src/lib/services/index.ts
@@ -23,3 +23,4 @@ export * from './valtimo-modal.service';
 export * from './page-header.service';
 export * from './cds-theme.service';
 export * from './page-subtitle.service';
+export * from './page-title.service';
