+++ projects/valtimo/components/src/public_api.ts
@@ -60,7 +60,7 @@ export * from './lib/components/page-header/page-header.module';
 
 export * from './lib/components/page-title/page-title.component';
 export * from './lib/components/page-title/page-title.module';
-export * from './lib/components/page-title/page-title.service';
+export * from './lib/services/page-title.service';
 
 export * from './lib/components/breadcrumb-navigation/breadcrumb-navigation.component';
 export * from './lib/components/breadcrumb-navigation/breadcrumb-navigation.module';
